+++ lib/index.js
@@ -1,30 +1,37 @@
+
+require('env2')('.env');
+// console.log(process.env);
+var google = require('googleapis');
+var OAuth2Client = google.auth.OAuth2;
+var CLIENT_ID = process.env.GOOGLE_CLIENT_ID;
+var CLIENT_SECRET = process.env.GOOGLE_CLIENT_SECRET;
+var REDIRECT_URL = 'http://localhost:8000/googleauth';
+var oauth2Client = new OAuth2Client(CLIENT_ID, CLIENT_SECRET, REDIRECT_URL);
+var plus = google.plus('v1');
+
 /**
- * register simply creates a /register POST route
+ * this plugin creates a /googleauth where google calls back to
  */
-exports.register = function (server, options, next) {
-  // console.log(server.info)
-  if(options && !options.fields) {
-    var msg = 'Please define required/optional fields see: http://git.io/vctLR';
-    console.log(msg)
-    return next(msg);
-  }
-  if(options && !options.handler){
-    var msg = 'Please specify a /register handler. see: http://git.io/vctLR';
-    console.log(msg)
-    return next(msg);
-  }
-  // expose /register route for any app that includes this plugin
-  server.route({
+exports.register = function googleauth (server, options, next) {
+  server.route([
+    {
+      method: 'GET',
+      path: '/',
+      handler: function(req, reply) {
+        var url = oauth2Client.generateAuthUrl({
+          access_type: 'offline', // will return a refresh token
+          scope: 'https://www.googleapis.com/auth/plus.profile.emails.read'
+          // can be a space-delimited string or an array of scopes
+        });
+        reply("<a href='" + url +"'>Click to Login!</a>" );
+      }
+    },
+    {
       method: '*',
-      path: '/register',
-      config: {
-        validate: {
-          payload : options.fields,
-          failAction: options.fail_action_handler || 'error'
-        }
-      },
-      handler: options.handler
-  });
+      path: '/googleauth',
+      handler: require('./google_oauth_handler.js')
+    }
+  ]);
 
   next(); // everything worked, continue booting the hapi server!
 };
